+++ iOSClient/Settings/Acknowledgements.rtf
@@ -316,4 +316,13 @@ The MIT License (MIT)\
 \
 Copyright (c) Wenchao Ding\
 ____________________________________________\
+\
+
+\f1\b DropDown
+\f0\b0 \
+\
+The MIT License (MIT)\
+\
+Copyright (c) Kevin Hirsch\
+____________________________________________\
 }
